refactor(provider-sdk): consolidate BPS fee utilities into provider-sdk
Moved shared BPS constants and conversion functions (computeBps,
bpsToRawFeeParams, isBpsPrecisionValid, assertBpsPrecision) to
provider-sdk as the single source of truth. sdk and svm-sdk now
import from provider-sdk instead of maintaining independent copies.
bpsToRawFeeParams takes maxInt and assumedMaxAmount as params to
support both EVM (MaxUint256, 10^36) and SVM (u64, 10^8).
