+++ typescript/sdk/src/fee/EvmTokenFeeReader.ts
@@ -33,13 +33,9 @@ import {
   getCrossCollateralRouterKeys,
   getEffectiveCrossCollateralDestinations,
 } from './crossCollateralUtils.js';
-import {
-  ASSUMED_MAX_AMOUNT_FOR_ZERO_SUPPLY,
-  BPS_PRECISION,
-  MAX_BPS,
-  assertBpsPrecision,
-  convertToBps,
-} from './utils.js';
+import { bpsToRawFeeParams } from '@hyperlane-xyz/provider-sdk/fee';
+
+import { ASSUMED_MAX_AMOUNT_FOR_ZERO_SUPPLY, convertToBps } from './utils.js';
 
 export type DerivedTokenFeeConfig = WithAddress<TokenFeeConfig>;
 type DerivedCrossCollateralFeeContracts = Record<
@@ -338,20 +334,10 @@ export class EvmTokenFeeReader extends HyperlaneReader {
   }
 
   convertFromBps(bps: number): FeeParameters {
-    if (!Number.isFinite(bps) || bps <= 0) {
-      throw new Error('bps must be > 0 to prevent division by zero');
-    }
-    assertBpsPrecision(bps);
-
-    const maxFee =
-      BigInt(constants.MaxUint256.toString()) /
-      ASSUMED_MAX_AMOUNT_FOR_ZERO_SUPPLY;
-    const scaledBps = BigInt(Math.round(bps * Number(BPS_PRECISION)));
-    const halfAmount = ((maxFee / 2n) * MAX_BPS * BPS_PRECISION) / scaledBps;
-
-    return {
-      maxFee,
-      halfAmount,
-    };
+    return bpsToRawFeeParams(
+      bps,
+      BigInt(constants.MaxUint256.toString()),
+      ASSUMED_MAX_AMOUNT_FOR_ZERO_SUPPLY,
+    );
   }
 }
